// src/client/Phm.MobileSp.Cms.v2/ClientApp/enums.ts
export enum FeedTypeEnum {
  Text = 0,
  Video = 1,
  Image = 2,
  Quiz = 3,
  Survey = 4,
  VideoLink = 5,
  ImageLink = 6,
  PdfLink = 7,
  Observation = 8,
  Paged = 9
}
export enum FeedCategoryEnum {
  Learning = 0,
  News = 1,
  General = 2,
  Campaign = 3,
  Announcement = 4,
  Article = 5
}
export enum MediaTypes {
  Text = 0,
  Font = 1,
  Image = 2,
  Video = 3,
  Icon = 4,
  File = 5,
}
export enum QuizQuestionTypeEnum {
  Single = 0,
  Multiple = 1
}
export enum SurveyQuestionTypeEnum {
  Single = 0,
  Multiple = 1
}
export enum AppTypeEnum {
  FISA = 0,
  CCESG = 1,
  MLearning = 2
}
export enum CopiedElementTypeEnum {
  CarModel = 0,
  CarSerie = 1,
  Technology = 2,
  Family = 3,
  CustomerQualification = 4,
  BodyColourSection = 5,
  CarOptionSection = 6,
  WalkAroundSteps = 7,
  CarSerieSpec = 8,
  MobileConfig = 9,
  ProductInformation = 10,
  LocalisationSet = 11,
  CompetitorsCarModel = 12,
  CompetitorsAssign = 13,
  VEDBandGroup = 14,
  Feed = 15,
  Competition = 16,
}
export enum PublishToTestObjectTypeEnum {
  CarModel = 0,
  CarSerie = 1,
  Technology = 2,
  Family = 3,
  CustomerQualification = 4,
  BodyColourSection = 5,
  CarOptionSection = 6,
  WalkAroundSteps = 7,
  CarSerieSpec = 8,
  MobileConfig = 9,
  ProductInformation = 10,
  BodyColour = 14,
  CarOption = 15,
  CarSerieFeature = 16,
  CustomerQualificationQuestion = 18,
  CustomerQualificationAnswer = 19,
  VEDBandGroup = 20,
}
export enum UploaderType {
  Any = 0,
  Image = 1,
  Video = 2,
  ImageAndVideo = 3,
  File = 4
}
export enum UserPointType {
  Unknown = 0,
  QuizFeed = 1,
  TextFeed = 2,
  VideoFeed = 3,
  SurveyFeed = 4
}
export enum BasePageFeedTypeEnum {
  Text = 0,
  MediaText = 1,
  MediaTabbedText = 2,
  Media = 3,
  TabbedText = 4
}
export enum BrandingElementType {
  Font = 0,
  Colour = 1,
  Icon = 2,
  CustomSelection = 3
}
export enum SecEntityTypeEnum {
  SecUser = 0,
  SecGroup = 1
}
export enum SecFeatureTypeEnum {
  Cms = 0,
  Mobile = 1
}